New shipment page #33
Display customer name
Display shipment details
Display list if tracking events with dates
Display field with list of tracking events that can be added to the list and "Add tracking event" button. After click on the button, new tracking event to be added to DB and to the list on the screen.